# scala_akka: update to latest versions and fix some settings