--- a/scala_akka_bench/Dockerfile
+++ b/scala_akka_bench/Dockerfile
@@ -1,7 +1,27 @@
 FROM hseeberger/scala-sbt:11.0.7_1.3.13_2.11.12 as BUILDER
 
 WORKDIR /app
-COPY scala_akka_bench /app
+
+# Make sure docker can cache a few static setup things
+
+# initialize only sbt
+RUN mkdir -p /app/project
+COPY scala_akka_bench/project/build.properties /app/project
+RUN sbt exit
+
+# initialize plugins
+COPY scala_akka_bench/project /app/project
+RUN sbt exit
+
+# initialize full build
+COPY scala_akka_bench/*.sbt /app
+# to initialize compiler
+RUN mkdir -p /app/src/main/scala/
+RUN touch /app/src/main/scala/Dummy.scala
+RUN sbt compile
+
+# Actually build project
+COPY scala_akka_bench/src /app/src
 COPY proto/helloworld/helloworld.proto /app/src/main/protobuf/helloworld.proto
 
 RUN sbt assembly