--- a/scala_akka_bench/build.sbt
+++ b/scala_akka_bench/build.sbt
@@ -2,18 +2,24 @@ name := "akka-grpc-quickstart-scala"
 
 version := "1.0"
 
-scalaVersion := "2.13.2"
+scalaVersion := "2.13.6"
 
 run / fork := true
 
-lazy val akkaVersion = "2.6.9"
-lazy val akkaGrpcVersion = "1.1.1"
+val akkaVersion = "2.6.15+8-dd289e7e-SNAPSHOT"
+val akkaHttpVersion = "10.2.5-M1"
 
 enablePlugins(AkkaGrpcPlugin)
+
+// to get latest versions
+resolvers += "Sonatype snapshots" at "https://oss.sonatype.org/content/repositories/snapshots"
 
 libraryDependencies ++= Seq(
   "ch.qos.logback" % "logback-classic" % "1.2.3",
   "com.typesafe.akka" %% "akka-actor-typed" % akkaVersion,
+  "com.typesafe.akka" %% "akka-http" % akkaHttpVersion,
+  "com.typesafe.akka" %% "akka-http-core" % akkaHttpVersion,
+  "com.typesafe.akka" %% "akka-parsing" % akkaHttpVersion,
   "com.typesafe.akka" %% "akka-stream" % akkaVersion,
   "com.typesafe.akka" %% "akka-discovery" % akkaVersion,
   "com.typesafe.akka" %% "akka-pki" % akkaVersion,
@@ -22,3 +28,12 @@ libraryDependencies ++= Seq(
   "com.typesafe.akka" %% "akka-stream-testkit" % akkaVersion % Test,
   "org.scalatest" %% "scalatest" % "3.1.1" % Test
 )
+
+// Akka and Google provided proto files seem to differ a bit so we need to choose
+// (doesn't seem to be important)
+assembly / assemblyMergeStrategy := {
+  case PathList(ps @ _*) if ps.last endsWith ".proto" => MergeStrategy.first
+  case x =>
+    val oldStrategy = (assembly / assemblyMergeStrategy).value
+    oldStrategy(x)
+}