--- a/scala_akka_bench/project/build.properties
+++ b/scala_akka_bench/project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.3.13
+sbt.version=1.5.4