--- a/scala_akka_bench/project/plugins.sbt
+++ b/scala_akka_bench/project/plugins.sbt
@@ -1,4 +1,4 @@
-addSbtPlugin("com.lightbend.akka.grpc" % "sbt-akka-grpc" % "1.1.1")
+addSbtPlugin("com.lightbend.akka.grpc" % "sbt-akka-grpc" % "2.1.0-M1")
 
 addSbtPlugin("com.lightbend.sbt" % "sbt-javaagent" % "0.1.5")