--- a/scala_akka_bench/src/main/resources/application.conf
+++ b/scala_akka_bench/src/main/resources/application.conf
@@ -1,0 +1,4 @@
+akka.http.server.max-connections = 1500
+akka.http.server.preview.enable-http2 = on
+akka.http.server.http2.min-collect-strict-entity-size = 1
+akka.actor.default-dispatcher.fork-join-executor.parallelism-max = ${GRPC_SERVER_CPUS}