@jtjeferreira 2021-07-01
with this PR:

```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_akka                  |   12475 |       79.54 ms |      134.81 ms |      164.16 ms |      287.94 ms |   71.08% |    164.35 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

before ( 0dd889b )

```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_akka                  |    3603 |      274.15 ms |      480.43 ms |      591.88 ms |         1.04 s |   97.92% |    241.43 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

:clap: